updated site, fixed backend and stuff

+++ wimf-processor/src/Post/Core/UseCases/SearchPost/SearchPostResponse.cs
@@ -1,10 +1,10 @@
 using System.Collections.Generic;
-
+#nullable disable
 namespace PostApi
 {
     public class SearchPostResponse
     {
-        public int Total { get; set; }
+        public long Total { get; set; }
 
         public IEnumerable<Post> Posts { get; set; }
     }
